feat: integrate i18n module into global module for enhanced localization support
